pyproject.toml
[project]
name = "storyevals"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "matplotlib>=3.10.5",
    "ruff>=0.12.9",
    "streamlit>=1.28.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.6.0",
    "mypy>=1.8.0",
    "types-requests",
    "types-PyYAML",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by ruff-format
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"



[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true



[dependency-groups]
dev = [
    "bandit>=1.8.6",
    "detect-secrets>=1.5.0",
    "mypy>=1.17.1",
    "pre-commit>=4.3.0",
    "types-pyyaml>=6.0.12.20250809",
    "types-requests>=2.32.4.20250809",
]
